modal/requirements.txt
llama-cpp-python[server]
fastapi
pydantic
modal